cb0617da-65f8-4a3c-adfe-c423abbbb9c9

+++ download/eaglercraft-26.1.2/sources/teavm/java/net/lax1dude/eaglercraft/v2_6/internal/PlatformRuntime.java
@@ -433,7 +433,7 @@ public class PlatformRuntime {
                 ArrayBuffer buffer = ArrayBuffer.create(arr.length);
                 Uint8Array view = Uint8Array.create(buffer);
                 for (int i = 0; i < arr.length; i++) {
-                        view.set(i, arr[i] & 0xFF);
+                        view.set(i, (short)(arr[i] & 0xFF));
                 }
                 return view;
         }
@@ -463,7 +463,7 @@ public class PlatformRuntime {
                 ArrayBuffer buffer = ArrayBuffer.create(arr.length);
                 Uint8Array view = Uint8Array.create(buffer);
                 for (int i = 0; i < arr.length; i++) {
-                        view.set(i, arr[i] & 0xFF);
+                        view.set(i, (short)(arr[i] & 0xFF));
                 }
                 return buffer;
         }
